Atualizações branch sabrina

+++ ListMEAPI/ListMEAPI/Repositories/UsuarioRepository.cs
@@ -1,5 +1,5 @@
-﻿using ListMEAPI.DTOs.Request;
-using ListMEAPI.Interfaces.Repositorios;
+﻿using ListMEAPI.DTOs.Request.Usuario;
+using ListMEAPI.Interfaces.Repositorios.Usuario;
 using ListMEAPI.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -18,18 +18,37 @@ namespace ListMEAPI.Repositories
             _context = ctx;
         }
 
-        public void Create(UsuarioModel usuario)
+        // POST
+        public dynamic Create(UsuarioModel usuario)
         {
+            var busca = _context.Usuarios.Any(a => a.Email == usuario.Email);
+            if (busca == true) {
+
+                return busca;
+
+            }               
+            
             _context.Add(usuario);
             _context.SaveChanges();
+
+            return null;
+
         }
 
         //DELETE
-        public void Delete(int id)
+        public bool Delete(int id)
         {
             var result = _context.Usuarios.Find(id);
-            _context.Usuarios.Remove(result);
-            _context.SaveChanges();
+            if (result != null)
+            {
+                _context.Usuarios.Remove(result);
+                _context.SaveChanges();
+                return true;
+            }
+            else
+            {
+                return false;
+            }
 
         }
 
@@ -64,38 +83,15 @@ namespace ListMEAPI.Repositories
                 return usuarioAntigo;
             }
             return usuarioAntigo;
-            
+
         }
-
-        //public ActionResult SubstituirUsuarioPelaId(int Id, UsuarioModel usuario)
-        //{
-        //    if (Id != usuario.Id_Usuario)
-        //    {
-        //        return BadRequest();
-        //    }
-        //    if (Id == usuario.Id_Usuario)
-        //    {
-        //        //Substitui valor da instância no banco de dados 
-        //        _listMEContext.Entry(usuario).State = EntityState.Modified;
-        //        _listMEContext.SaveChanges();
-
-            //        return NoContent();
-            //    }
-
-            if (id == usuario.Id_Usuario)
-            {
-
-                _context.Entry(usuario).State = EntityState.Modified;
-                _context.SaveChanges();
-            }
-        }
-
-
 
         public void Save()
         {
             _context.SaveChanges();
         }
+
     }
 }
+
